add token info for named types, handle tokens with same names, refactoring
