+++ semantics/src/main/scala/aqua/semantics/rules/abilities/AbilitiesInterpreter.scala
@@ -56,11 +56,6 @@ class AbilitiesInterpreter[S[_], X](using
       case false =>
         for {
           _ <- modify(_.defineService(name, defaultId))
-          _ <- locations.addTokenWithFields(
-            name.value,
-            name,
-            arrowDefs.toNel.toList
-          )
         } yield true
     }
