+++ semantics/src/main/scala/aqua/semantics/rules/locations/LocationsAlgebra.scala
@@ -6,7 +6,7 @@ case class TokenInfo[S[_]](token: Token[S], `type`: Type)
 
 trait LocationsAlgebra[S[_], Alg[_]] {
   def addToken(name: String, tokenInfo: TokenInfo[S]): Alg[Unit]
-  def addTokenWithFields(name: String, token: Token[S], fields: List[(String, Token[S])]): Alg[Unit]
+  def addTokenWithFields(name: String, token: TokenInfo[S], fields: List[(String, TokenInfo[S])]): Alg[Unit]
 
   def pointTokenWithFieldLocation(typeName: String, typeToken: Token[S], fieldName: String, token: Token[S]): Alg[Unit]
   def pointFieldLocation(typeName: String, fieldName: String, token: Token[S]): Alg[Unit]
